#64 rename database module
