+++ backend/store/json_store.go
@@ -1,4 +1,4 @@
-package database
+package store
 
 import (
 	"encoding/json"
